[HTTP] refactored prepare stats request to common http code.
