+++ pkg/http/common.go
@@ -1,7 +1,11 @@
 package http
 
 import (
+	"fmt"
+	"strconv"
+
 	"github.com/gocql/gocql"
+	"github.com/surahman/mcq-platform/pkg/auth"
 	"github.com/surahman/mcq-platform/pkg/cassandra"
 	"github.com/surahman/mcq-platform/pkg/model/cassandra"
 	"github.com/surahman/mcq-platform/pkg/redis"
@@ -34,4 +38,25 @@ func GetQuiz(quizId gocql.UUID, db cassandra.Cassandra, cache redis.Redis) (*mod
 	}
 
 	return &quiz, nil
+}
+
+// PrepareStatsRequest will prepare the paged statistics request for the database query.
+func PrepareStatsRequest(auth auth.Auth, quizId gocql.UUID, cursor string, size string) (req *model_cassandra.StatsRequest, err error) {
+	req = &model_cassandra.StatsRequest{QuizID: quizId}
+
+	if req.PageSize, err = strconv.Atoi(size); err != nil {
+		return nil, fmt.Errorf("failed to convert page size: %s", err.Error())
+	}
+	if req.PageSize < 1 {
+		req.PageSize = 10
+	}
+
+	// Null cursor must be set if there was no cursor in the URI.
+	if len(cursor) != 0 {
+		if req.PageCursor, err = auth.DecryptFromString(cursor); err != nil {
+			return nil, fmt.Errorf("failed to decrypt page cursor: %s", err.Error())
+		}
+	}
+
+	return
 }
